requirements.txt
torch>=2.0.0
transformers>=4.30.0
pytorch-lightning>=2.0.0
gradio>=4.0.0
wandb>=0.15.0
datasets>=2.12.0
pyyaml>=6.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface-hub>=0.15.0
flask>=2.0.0 